requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-groq==0.2.0
langchain-core==0.3.0
langchain-huggingface==0.1.0
langchain-text-splitters==0.3.0
langsmith==0.1.147
groq==0.11.0
faiss-cpu==1.8.0
sentence-transformers==3.1.1
rank-bm25==0.2.2
numpy==1.26.4
Pillow==10.4.0
requests==2.32.3
streamlit==1.38.0
python-dotenv==1.0.1
pydantic==2.9.2
torch==2.4.0
transformers==4.44.0
tokenizers==0.19.1